Improve operator performance
This makes iterating over the NA BitArray more efficient and removes
bounds checks in places where they make a difference, which gives a
significant (~2x) performance boost. In the absence of NAs, for +
DataArrays are within 30% of performance for Arrays. For .* we are
basically equivalent. For .== we are ~20% faster, probably because .==
on Arrays returns a BitArray.