Fix #76
Note that I made AbstractDataArray inherit from StoredArray instead
of AbstractArray, since both DataArray and PooledDataArray are
StoredArrays. DataArray should probably inherit from DenseArray, and
the inheritance should probably be on those types and not
AbstractDataArray, but we don't have multiple inheritance, so this
seems like the way to go for now.