WIP: Internal API changes due to 10525
There were a few internal APIs that I changed in JuliaLang/julia#10525.  This is a start at fixing some.  I think I got everything except a few assumptions about how Base.checkbounds will work with NAType, so this fails in a few places where you expect checkbounds to throw an NAException -- it is now a MethodError.

There is a lot more that you could do to adjust to a post-10525 world.  Let me know if you have any questions along the way.