Fix conversion between Array{Union{T, Null}} and DataArray
Previously, convert() would create a DataArray{Union{T, Null}} containing
nulls which were not considered as missing, giving weird results in particular
with equality tests.

Constructing a DataArray{T} from an Array{Union{T, Null}} is tricky because
the field must be an Array{T}, yet conversion will fail in the presence of
nulls. Therefore, we need to allocate a copy and leave the null entries
uninitialized.