Restrict iterators to DataArrays
Optimizations need to rely on the particular fields of DataArray.
This requires moving code to dataarray.jl since the type had not
been defined in abstractdataarray.jl.

Rather than adding specialized iterators for PooledDataArray, which
is going to be deprecated, rely on the generic iterators provided
by Nulls, which are already faster than the current ones.